Move to arbitrarily-sized integers
- frontend makes BigInt
- ref backend uses BigInt
- truffle backend uses Java BigInteger

+++ commons/src/main/scala/org/bitbucket/inkytonik/cooma/Backend.scala
@@ -14,7 +14,7 @@ trait Backend {
 
     type Value
     def funV(k : String, x : String, body : Term) : Value
-    def intV(i : Int) : Value
+    def intV(i : BigInt) : Value
     def prmV(p : Primitive, xs : Vector[String]) : Value
     def rowV(fs : Vector[FieldValue]) : Value
     def strV(s : String) : Value
